+++ R/cmap.r
@@ -53,7 +53,17 @@ cmap <- function(.x, .f, ...) {
 #' @rdname cmap
 #' @importFrom future getGlobalsAndPackages
 #' @export
-cmap.disk.frame <- function(.x, .f, ..., outdir = NULL, keep = NULL, chunks = nchunks(.x), compress = 50, lazy = TRUE, overwrite = FALSE, vars_and_pkgs = future::getGlobalsAndPackages(.f, envir = parent.frame()), .progress = TRUE) {
+cmap.disk.frame <- function(
+                    .x, 
+                    .f, 
+                    ..., 
+                    outdir = NULL, 
+                    keep = NULL, 
+                    chunks = nchunks(.x), 
+                    compress = 50, 
+                    lazy = TRUE, 
+                    overwrite = FALSE, 
+                    vars_and_pkgs = future::getGlobalsAndPackages(.f, envir = parent.frame()), .progress = TRUE) {
   .f = purrr::as_mapper(.f)
   if(lazy) {
     attr(.x, "lazyfn") = 
@@ -93,12 +103,11 @@ cmap.disk.frame <- function(.x, .f, ..., outdir = NULL, keep = NULL, chunks = nc
   dotdotdot = list(...)
   
   res = future.apply::future_lapply(1:length(files), function(ii, ...) {
-    #res = lapply(1:length(files), function(ii) {
+  #res = lapply(1:length(files), function(ii) {
     ds = disk.frame::get_chunk(.x, cid[ii], keep=keep_future, full.names = TRUE)
     
     res = .f(ds, ...)
-    
-    #res = do.call(.f, c(ds, dotdotdot))
+    # res = do.call(.f, c(ds, dotdotdot))
     
     if(!is.null(outdir)) {
       if(nrow(res) == 0) {
@@ -110,7 +119,9 @@ cmap.disk.frame <- function(.x, .f, ..., outdir = NULL, keep = NULL, chunks = nc
     } else {
       return(res)
     }
-  }, ...)
+  }, ..., 
+  future.seed=TRUE # to get rid of the error TODO investigate making this better
+  )
   
   if(!is.null(outdir)) {
     return(disk.frame(outdir))
@@ -183,7 +194,7 @@ cimap.disk.frame <- function(.x, .f, outdir = NULL, keep = NULL, chunks = nchunk
     } else {
       return(res)
     }
-  })
+  }, future.seed = TRUE)
   
   if(!is.null(outdir)) {
     return(disk.frame(outdir))