+++ R/dplyr_verbs.r
@@ -31,12 +31,6 @@ select.disk.frame <- function(.data, ...) {
   }, lazy = TRUE)
 }
 
-#' Kept for backwards-compatibility to be removed in 0.3
-#' @export
-create_dplyr_mapper = function() {
-  stop("create_dplyr_mapper has been deprecated. Please use create_chunk_mapper instead")
-}
-
 #' @export
 #' @rdname dplyr_verbs
 rename.disk.frame <- create_chunk_mapper(dplyr::rename)
@@ -72,27 +66,20 @@ arrange.disk.frame =create_chunk_mapper(dplyr::arrange, warning_msg="`arrange.di
 chunk_arrange <- create_chunk_mapper(dplyr::arrange)
 
 
-#' @export
-#' @importFrom dplyr tally
-#' @rdname dplyr_verbs
-tally.disk.frame <- create_chunk_mapper(dplyr::tally)
+# TODO family is not required is group-by
+# TODO alot of these .disk.frame functions are not generic
 
 
 #' @export
-#' @importFrom dplyr count
+#' @importFrom dplyr add_count
 #' @rdname dplyr_verbs
-count.disk.frame <- create_chunk_mapper(dplyr::count)
+add_count.disk.frame <- create_chunk_mapper(dplyr::add_count)
 
-#' #' @export
-#' #' @importFrom dplyr add_count
-#' #' @rdname dplyr_verbs
-#' add_count.disk.frame <- create_chunk_mapper(dplyr::add_count)
-#' 
-#' 
-#' #' @export
-#' #' @importFrom dplyr add_tally
-#' #' @rdname dplyr_verbs
-#' add_tally.disk.frame <- create_chunk_mapper(dplyr::add_tally)
+
+#' @export
+#' @importFrom dplyr add_tally
+#' @rdname dplyr_verbs
+add_tally.disk.frame <- create_chunk_mapper(dplyr::add_tally)
 
 
 #' @export