+++ tutorials/fannie_mae/5b_one_var.r
@@ -1,4 +1,4 @@
-source("inst/fannie_mae/0_setup.r")
+source("tutorials/fannie_mae/00_setup.r")
 library(disk.frame)
 
 acqall1 = disk.frame(file.path(outpath, "appl_mdl_data"))